Rename actor to worker
